[devmgr] Add platform bus support
The platform-bus is a special driver that is able to instantiate devices
based on information in the MDI.
The devmgr special cases this driver and launches it in a separate devhost
and passes it a copy of the MDI data from the bootdata passed in from the kernel.

When the platform-bus driver binds, it creates devices corresponding to
the "device" nodes contained in the top level "platform-bus" node in the MDI.
Other drivers can bind to these nodes based on a vid/pid/did protocol.
These devices can optionally have children specified in the MDI.
After the drivers for these devices bind, they can call the
platform_device_add_children() helper function to instantiate their
devices from the MDI.

Currently platform-bus only supports instantiating a tree of devices from the MDI.
In upcoming changes we will add support for specifying IRQs, MMIO ranges, etc
in the MDI and passing these to drivers via resources.

Change-Id: Icf57e486b4c93b1cf68910886f57744a5392cc7d